Extract base ui components
