+++ app/src/debug/kotlin/io/sweers/catchup/ui/activity/UiModule.kt
@@ -17,9 +17,9 @@ package io.sweers.catchup.ui.activity
 
 import dagger.Module
 import dagger.Provides
+import io.sweers.catchup.base.ui.ViewContainer
 import io.sweers.catchup.injection.scopes.PerActivity
 import io.sweers.catchup.ui.DebugViewContainer
-import io.sweers.catchup.ui.ViewContainer
 import io.sweers.catchup.ui.bugreport.BugReportModule
 
 @Module(includes = [BugReportModule::class])
