+++ app/src/main/kotlin/io/sweers/catchup/util/UiUtil.kt
@@ -30,6 +30,7 @@ import androidx.annotation.ColorInt
 import androidx.annotation.FloatRange
 import androidx.interpolator.view.animation.FastOutSlowInInterpolator
 import androidx.palette.graphics.Palette
+import io.sweers.catchup.base.ui.ColorUtils
 
 object UiUtil {
