+++ libraries/base-ui/src/main/kotlin/io/sweers/catchup/base/ui/BaseActivity.kt
@@ -1,11 +1,11 @@
 /*
- * Copyright (C) 2019. Zac Sweers
+ * Copyright (c) 2019 Zac Sweers
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ *   http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.sweers.catchup.ui.base
+package io.sweers.catchup.base.ui
 
 import android.annotation.SuppressLint
 import android.os.Bundle
@@ -25,17 +25,13 @@ import androidx.core.app.NavUtils
 import com.jakewharton.rxrelay2.BehaviorRelay
 import com.uber.autodispose.lifecycle.CorrespondingEventsFunction
 import com.uber.autodispose.lifecycle.LifecycleScopeProvider
-import dagger.android.AndroidInjection
 import io.reactivex.Observable
-import io.sweers.catchup.CatchUpPreferences
-import io.sweers.catchup.ui.ViewContainer
-import io.sweers.catchup.ui.base.ActivityEvent.CREATE
-import io.sweers.catchup.ui.base.ActivityEvent.DESTROY
-import io.sweers.catchup.ui.base.ActivityEvent.PAUSE
-import io.sweers.catchup.ui.base.ActivityEvent.RESUME
-import io.sweers.catchup.ui.base.ActivityEvent.START
-import io.sweers.catchup.ui.base.ActivityEvent.STOP
-import io.sweers.catchup.util.updateNavBarColor
+import io.sweers.catchup.base.ui.ActivityEvent.CREATE
+import io.sweers.catchup.base.ui.ActivityEvent.DESTROY
+import io.sweers.catchup.base.ui.ActivityEvent.PAUSE
+import io.sweers.catchup.base.ui.ActivityEvent.RESUME
+import io.sweers.catchup.base.ui.ActivityEvent.START
+import io.sweers.catchup.base.ui.ActivityEvent.STOP
 import javax.inject.Inject
 
 abstract class BaseActivity : AppCompatActivity(), LifecycleScopeProvider<ActivityEvent> {
@@ -116,7 +112,7 @@ abstract class BaseActivity : AppCompatActivity(), LifecycleScopeProvider<Activi
   @Inject
   protected lateinit var viewContainer: ViewContainer
   @Inject
-  protected lateinit var catchUpPreferences: CatchUpPreferences
+  protected lateinit var uiPreferences: UiPreferences
 
   @CheckResult
   override fun lifecycle(): Observable<ActivityEvent> {
@@ -133,17 +129,8 @@ abstract class BaseActivity : AppCompatActivity(), LifecycleScopeProvider<Activi
 
   @CallSuper
   override fun onCreate(savedInstanceState: Bundle?) {
-    AndroidInjection.inject(this)
-    setFragmentFactory()
     super.onCreate(savedInstanceState)
     lifecycleRelay.accept(CREATE)
-  }
-
-  /**
-   * Exists as a hook to allow subclasses to inject a FragmentFactory before
-   * super.onCreate() is called.
-   */
-  open fun setFragmentFactory() {
   }
 
   @CallSuper
@@ -167,7 +154,7 @@ abstract class BaseActivity : AppCompatActivity(), LifecycleScopeProvider<Activi
 
   override fun onAttachedToWindow() {
     super.onAttachedToWindow()
-    updateNavBarColor(catchUpPreferences = catchUpPreferences)
+    updateNavBarColor(uiPreferences = uiPreferences)
   }
 
   @CallSuper
