+++ libraries/base-ui/src/main/kotlin/io/sweers/catchup/base/ui/ViewContainer.kt
@@ -1,11 +1,11 @@
 /*
- * Copyright (C) 2019. Zac Sweers
+ * Copyright (c) 2019 Zac Sweers
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ *   http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -13,10 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.sweers.catchup.ui
+package io.sweers.catchup.base.ui
 
 import android.view.ViewGroup
-import io.sweers.catchup.ui.base.BaseActivity
 
 /**
  * An indirection which allows controlling the root container used for each activity.
